Youtube Chatbot using LangChain/requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface
langchain-google-genai
youtube-transcript-api
faiss-cpu
python-dotenv
